refactor money summary
